requirements.txt
numpy==1.26.4
transformers==4.46.3
accelerate==1.1.1
# deepspeed==0.15.4
# peft==0.13.2
PyYAML==6.0.2
datasets==3.1.0
sentencepiece==0.2.0
tyro==0.6.0
tqdm==4.67.1
wandb==0.18.7
omegaconf==2.3.0
pillow==11.0.0

# trl==0.13.0